backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
motor==3.6.0
pydantic==2.7.4
pydantic-settings==2.3.4
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
httpx==0.27.0
slowapi==0.1.9
pytest==7.4.4
pytest-asyncio==0.23.3
